Enable unsafe narrowing checks in cross-platform WebNN code
These warnings are useful to detect potential issues. This requires
adding explicit integer casts to command_buffer_id.h and vulkan_image.h
as well since these GPU headers are included by WebNN.

To avoid adding these checks to platform-specific WebNN code the build
targets have been split however this requires adding some "nogncheck"
annotations to temporarily ignore the circular dependency this creates
between the new and existing targets.

Bug: 493652457
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7788132
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1622028}

+++ gpu/ipc/common/command_buffer_id.h
@@ -18,7 +18,8 @@ enum class GpuChannelReservedRoutes : int32_t {
 inline CommandBufferId CommandBufferIdFromChannelAndRoute(int channel_id,
                                                           int32_t route_id) {
   return CommandBufferId::FromUnsafeValue(
-      (static_cast<uint64_t>(channel_id) << 32) | route_id);
+      (static_cast<uint64_t>(channel_id) << 32) |
+      static_cast<uint32_t>(route_id));
 }
 
 inline int ChannelIdFromCommandBufferId(
@@ -28,7 +29,7 @@ inline int ChannelIdFromCommandBufferId(
 
 inline int32_t RouteIdFromCommandBufferId(
     gpu::CommandBufferId command_buffer_id) {
-  return 0xffffffff & command_buffer_id.GetUnsafeValue();
+  return static_cast<int32_t>(0xffffffff & command_buffer_id.GetUnsafeValue());
 }
 
 }  // namespace gpu
